backend/requirements.txt
# F1 AI Race Engineer — Dependencies
# Install: pip install -r requirements.txt

# ── Core data pipeline ──
fastf1>=3.3.0
pandas>=2.2.0
numpy>=1.26.0

# ── AI / Embeddings / Vector search ──
# google-genai>=1.0.0
groq>=0.9.0
sentence-transformers>=3.0.0
torch>=2.0.0
faiss-cpu>=1.7.4

# ── Web server ──
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=22.0.0

# ── Config ──
python-dotenv>=1.0.0

# ── Schema validation ──
pydantic>=2.0.0

# ── MCP (Model Context Protocol) ──
mcp>=1.0.0
